+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [unreleased] - 2023-05-23
 
 ### Fixed
-* Fix a 'callback_queue' exception in Ruby 3.1.1 and higher.
+* Fix a 'callback_queue' exception in Ruby 3.1.1 by Vadim Kononov(@vkononov)
 
 ## [v0.14.0] - 2020-09-12