+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+### Fixed
+* Fix defining object finalizer in Ruby 3.1 to reference a UUID string
+  by Vadim Kononov(@vkononov)
+
 ## [v0.14.0] - 2020-09-12
 
 ### Added